+++ src/Network/Network.Test/ScenarioTests/AzureFirewallPolicyTests.cs
@@ -57,5 +57,13 @@ namespace Commands.Network.Test.ScenarioTests
         {
             TestRunner.RunTestScript("Test-AzureFirewallPolicyCRUDWithNetworkRuleDestinationFQDNs");
         }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        [Trait(Category.Owner, NrpTeamAlias.azurefirewall)]
+        public void TestAzureFirewallPolicyWithIpGroups()
+        {
+            TestRunner.RunTestScript("Test-AzureFirewallPolicyWithIpGroups");
+        }
     }
 }